requirements.txt
customtkinter~=5.2.2
pillow~=12.1.1
pdfplumber~=0.11.9
reportlab~=4.4.10